Move server code into server/ subdir for consistency
