+++ server/.eslintignore
@@ -1,0 +1,2 @@
+*.min.js
+coverage
