+++ server/package.json
@@ -1,0 +1,83 @@
+{
+  "name": "probot-webhooks",
+  "version": "0.0.3",
+  "description": "",
+  "author": "Jason Etcovitch <jasonetco@gmail.com> (https://github.com/JasonEtco)",
+  "license": "ISC",
+  "repository": "https://github.com/probot/webhooks.git",
+  "scripts": {
+    "start": "node ./index.js",
+    "start-dev": "concurrently \"nodemon --ignore src/ ./index.js\" \"webpack -w\"",
+    "build": "webpack -p",
+    "test": "jest --coverage && eslint",
+    "postinstall": "npm run build"
+  },
+  "dependencies": {
+    "autoprefixer": "^7.1.6",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "connect-sse": "^1.2.0",
+    "copy-to-clipboard": "^3.0.8",
+    "copy-webpack-plugin": "^4.2.0",
+    "crypto": "^1.0.1",
+    "css-loader": "^0.28.7",
+    "eventsource": "^1.0.5",
+    "express": "^4.16.2",
+    "express-sslify": "^1.2.0",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "get-value": "^2.0.6",
+    "glob-all": "^3.1.0",
+    "helmet": "^3.9.0",
+    "html-webpack-plugin": "^2.30.1",
+    "moment": "^2.19.1",
+    "moment-timezone": "^0.5.14",
+    "node-sass": "^4.5.3",
+    "postcss-loader": "^2.0.8",
+    "primer-css": "^9.6.0",
+    "prop-types": "^15.6.0",
+    "purify-css": "^1.2.5",
+    "purifycss-webpack": "^0.7.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-json-view": "^1.13.2",
+    "react-octicons": "^0.2.0",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.19.0",
+    "webpack": "^3.8.1"
+  },
+  "devDependencies": {
+    "babel-jest": "^21.2.0",
+    "concurrently": "^3.5.0",
+    "enzyme": "^3.2.0",
+    "enzyme-adapter-react-16": "^1.1.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-config-standard-react": "^5.0.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.5.1",
+    "eslint-plugin-standard": "^3.0.1",
+    "jest": "^21.2.1",
+    "nodemon": "^1.12.1",
+    "raf": "^3.4.0",
+    "react-test-renderer": "^16.2.0",
+    "standard": "^10.0.3",
+    "supertest": "^3.0.0"
+  },
+  "engines": {
+    "node": "8.9.1"
+  },
+  "jest": {
+    "setupFiles": [
+      "./tests/setup.js"
+    ],
+    "testPathIgnorePatterns": [
+      "/node_modules/"
+    ]
+  }
+}
