+++ server/src/components/EventDescription.js
