+++ server/src/style.scss
