+++ server/src/styles/header-anim.scss
