+++ server/src/styles/main-anim.scss
